requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
pandas==2.2.1
requests==2.31.0
pydantic==2.6.4
python-dotenv==1.0.1
